requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# AI Medical Assistant – Requirements
# Install with: pip install -r requirements.txt
# ──────────────────────────────────────────────────────────────────────────────

# ── Web Framework ──────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── LLM Provider ──────────────────────────────────────────────────────────────
anthropic>=0.28.0

# ── LangChain Core & Community ────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0

# ── Vector Database ───────────────────────────────────────────────────────────
faiss-cpu>=1.8.0           # CPU-only FAISS (use faiss-gpu if you have a CUDA GPU)

# ── Embeddings (local, no API key needed) ────────────────────────────────────
sentence-transformers>=2.7.0
torch>=2.0.0               # Required by sentence-transformers

# ── PDF Loading ───────────────────────────────────────────────────────────────
pypdf>=4.0.0               # For loading PDF documents

# ── Environment Variables ─────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy>=1.26.0
